chapter 7 complete
